Refactor search bar and update layout, styling
